24 game
program that randomly chooses and displays four digits, each from one to nine, with repetitions allowed. The program prompts for the player to enter an arithmetic expression using just those, and all of those four digits, used exactly once each. The program checks then evaluate the expression. The goal is for the player to enter an expression that evaluates to 24.